cleanup of table view
