+++ src/ui/src/pages/HomePage.tsx
@@ -10,6 +10,7 @@ import type { BadgeVariant, ShaveItem } from "../types";
 import HeadingTag from "@/components/ui/heading-tag";
 import { ToggleGroup, ToggleGroupItem } from "@/components/ui/toggle-group";
 import { Table, TableBody, TableCell, TableHead, TableHeader, TableRow } from "@/components/ui/table";
+import { getYouTubeThumbnail, timeAgo } from "@/lib/utils";
 
 const NO_SHAVES_STEPS: string[] = ["Record screen and describe the issue", "AI transcribes and analyzes content", "Receive a structured work item ready to send"]
 
@@ -115,18 +116,6 @@ const ShaveCards = ({ shaves }: { shaves: ShaveItem[] }) => {
   )
 }
 
-function timeAgo(date: Date): string {
-  const seconds = Math.floor((Date.now() - date.getTime()) / 1000);
-  if (seconds < 60) return "just now";
-  const minutes = Math.floor(seconds / 60);
-  if (minutes < 60) return `${minutes} min ago`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours} hrs ago`;
-  const days = Math.floor(hours / 24);
-  if (days < 30) return `${days} days ago`;
-  return date.toLocaleDateString();
-}
-
 const ShaveStatusAction = ({ shave }: { shave: ShaveItem }) => {
   if (shave.shaveStatus === "Processing") {
     return (
@@ -162,6 +151,7 @@ const ShaveStatusAction = ({ shave }: { shave: ShaveItem }) => {
 };
 
 const ShaveTable = ({ shaves }: { shaves: ShaveItem[] }) => {
+  console.log(shaves)
   return (
     <Table>
       <TableHeader>
@@ -177,13 +167,23 @@ const ShaveTable = ({ shaves }: { shaves: ShaveItem[] }) => {
         {shaves.map((shave) => (
           <TableRow key={shave.id}>
             <TableCell>
-              <div
-                className="w-[100px] h-[56px] rounded bg-black/40 border border-white/10 flex items-center justify-center cursor-pointer"
-                onClick={() => shave.videoEmbedUrl && window.open(shave.videoEmbedUrl, "_blank")}
-                title={shave.videoEmbedUrl ? "Open video" : "No video"}
-              >
-                <Video className="h-5 w-5 text-muted-foreground" />
-              </div>
+              {(() => {
+                const thumbnail = shave.videoEmbedUrl ? getYouTubeThumbnail(shave.videoEmbedUrl) : null;
+                const videoUrl = shave.videoEmbedUrl?.replace("embed/", "watch?v=") || null;
+                const Wrapper = videoUrl ? "a" : "div";
+                return (
+                  <Wrapper
+                    className={`w-[100px] h-[56px] rounded bg-black/40 border border-white/10 flex items-center justify-center overflow-hidden ${videoUrl ? "cursor-pointer" : "cursor-default opacity-50"}`}
+                    {...(videoUrl ? { href: videoUrl, target: "_blank", rel: "noopener noreferrer", title: "Open video" } : {})}
+                  >
+                    {thumbnail ? (
+                      <img src={thumbnail} alt={shave.title} className="w-full h-full object-cover" />
+                    ) : (
+                      <Video className="h-5 w-5 text-muted-foreground" />
+                    )}
+                  </Wrapper>
+                );
+              })()}
             </TableCell>
             <TableCell className="font-medium max-w-[400px] truncate" title={shave.title}>
               {shave.title}
